+++ R/read.R
@@ -323,7 +323,7 @@ abbreviate_shapefile_names = function(x) {
 #'   is missing, the \code{basename} of \code{dsn} is taken.
 #' @param driver character; name of driver to be used; if missing and \code{dsn} is not a Database Connection, a driver name is guessed from \code{dsn};
 #' \code{st_drivers()} returns the drivers that are available with their properties; links to full driver documentation
-#' are found at \url{https://gdal.org/ogr_formats.html}.
+#' are found at \url{https://gdal.org/drivers/vector/index.html}
 #' @param ... other arguments passed to \link{dbWriteTable} when \code{dsn} is a
 #' Database Connection
 #' @param dataset_options character; driver dependent dataset creation options;